requirements.txt
pdfplumber
pandas
pyarrow
duckdb
streamlit
matplotlib
openai
python-dotenv
openpyxl
tabulate
langfuse
duckdb
